:recycle: Better files/folder organization and refactored the url code
This commits provides a better file organization in the src/network
contains the code for url, http client and http response. And for the
refactoring part the URL now only parses URLs and handles the connection
with the web server via socket the HttpClient is in charge for making
request, following redirects and displaying files/data. And for the last
there is the HttpResponse that parses responses and headers.